feat: add omnicode and native tool manager foundation
Add separate omnicode and omniproxy binaries, refactor native agent tools into focused files, expand the built-in tool set, and add install scripts for Unix and Windows.
