+++ cmd/omnicode/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "omnillm/internal/omnicode"
+
+func main() {
+	omnicode.Run()
+}
