+++ internal/omnicode/config.go
@@ -1,0 +1,8 @@
+package omnicode
+
+// Config is a placeholder for OmniCode-specific runtime configuration.
+type Config struct {
+	ServerURL string
+	APIKey    string
+	Output    string
+}
