+++ internal/tools/errors.go
@@ -1,0 +1,5 @@
+package tools
+
+import "errors"
+
+var errStopWalk = errors.New("stop walk")
