+++ internal/tools/strings.go
@@ -1,0 +1,7 @@
+package tools
+
+import "strings"
+
+func joinLines(lines []string) string {
+	return strings.Join(lines, "\n")
+}
